requirements.txt
PyYAML
jieba
scipy
